Make binary placeholders for a drat repository

The goal of drat is to let you create a source repository. Supporting the inclusion of binary packages is beyond the scope of these tools. However, if you do not have the directory structure for binaries, you will get some annoying warnings each time you install packages. This function will help you get rid of those messages.

make_drat_bin_placeholders()

Examples

  make_drat_bin_placeholders()

#> These placeholder-files will be created relative to project-root:
#>   '/Users/ijlyttle/Documents/git/github/public_me/ghentr/'
#>   ❯ 'bin/windows/contrib/3.1/PACKAGES'
#>   ❯ 'bin/macosx/mavericks/contrib/3.1/PACKAGES'
#>   ❯ 'bin/macosx/el-capitan/contrib/3.1/PACKAGES'
#>   ❯ 'bin/windows/contrib/3.2/PACKAGES'
#>   ❯ 'bin/macosx/mavericks/contrib/3.2/PACKAGES'
#>   ❯ 'bin/macosx/el-capitan/contrib/3.2/PACKAGES'
#>   ❯ 'bin/windows/contrib/3.3/PACKAGES'
#>   ❯ 'bin/macosx/mavericks/contrib/3.3/PACKAGES'
#>   ❯ 'bin/macosx/el-capitan/contrib/3.3/PACKAGES'
#>   ❯ 'bin/windows/contrib/3.4/PACKAGES'
#>   ❯ 'bin/macosx/mavericks/contrib/3.4/PACKAGES'
#>   ❯ 'bin/macosx/el-capitan/contrib/3.4/PACKAGES'
#> 
#> Proceed?
#> 1: Yes
#> 2: No
#> 3: Hell no
#>