Use GitHub Enterprise components in your package

These functions are used to write function files for your package.

use_github_enterprise(host, suffix, name)

use_install_ghe(host, suffix, name)

use_ghe_pat(suffix, name)

use_use_ghe(host, suffix, name)

Arguments

host
character hostname for GHE instance

suffix
character lowercase suffix to use for function names

name
character title-case name of the GHE instance

Value

invisible(NULL), called for side effects

Details

The function use_github_enterprise() calls all the templating functions:

use_install_ghe()

Writes a file to the current project, R/utils-install_ghe.R, containing a function install_github_foo() that wraps devtools::install_github(), using the host argument of this function.

use_ghe_pat()

Writes a file to the current project, R/utils-ghe_pat.R, containing a function github_foo_pat() that parallels devtools::github_pat(). This new function will look for an environment variable named GITHUB_FOO_PAT.

use_use_ghe()

Writes a file, utils-use_ghe.R, that contains a function use_github_foo() that wraps usethis::use_github(), using the host argument of this function.

For each of the functions described above, the foo in the function name is the value of the suffix argument from this function. The name argument from this function is interpolated into the documentation of each of written functions.

Examples

# NOT RUN {
  use_github_enterprise(
    host = "github.acme-corp.com/api/v3",
    suffix = "acme",
    name = "Acme Corporation"
  )
# }